+++ docs/examples/source/octocat_say.py
@@ -1,0 +1,12 @@
+import github3
+
+print("Hey Octocat!")
+print(github3.octocat("Hey Ian!"))
+print("What do you think about github3.py?")
+print(github3.octocat("github3.py rocks!"))
+print("Thanks Octocat, that means a lot coming from you.")
+print("FIN.")
+print("""Epilog:
+    The preceding conversation was entirely fictional. If you didn't realize
+    that, you need to get out more.
+""")